# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API RESTFUL Laravel 8.0",
        "description": "API documentation project",
        "contact": {
            "email": "ricardokovalskicruz@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/cakes": {
            "get": {
                "tags": [
                    "Cakes"
                ],
                "summary": "List of cakes",
                "description": "Return a list of cakes",
                "operationId": "2031c6fd92e54e697134adef02a03696",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/machines": {
            "post": {
                "tags": [
                    "Cakes"
                ],
                "summary": "Store a cake",
                "description": "Return a cake",
                "operationId": "a4b9d6dec7801e8b6c190aeb88064063",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "weight",
                        "in": "query",
                        "description": "Weight",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Value",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store cake"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/cakes/{id}": {
            "get": {
                "tags": [
                    "Cakes"
                ],
                "summary": "Show a cake",
                "description": "Return a cake",
                "operationId": "a94f7060c2c2512ecfd9bdc0a8864974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cake",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cakes"
                ],
                "summary": "Update a cake",
                "description": "Update a cake",
                "operationId": "778fd92249a1a41196239fdc9bd1c8ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cake to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "weight",
                        "in": "query",
                        "description": "Weight",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "description": "Value",
                        "required": true,
                        "schema": {
                            "type": "float"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update cake"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cakes"
                ],
                "summary": "Delete a cake",
                "description": "Delete a cake",
                "operationId": "49571687632ad0311a8f9b8dce0a706c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cake to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/v1/cakes/{id}/subscribe": {
            "post": {
                "tags": [
                    "Interested"
                ],
                "summary": "Subscribe a customer in cake",
                "description": "Subscribe a customer in cake",
                "operationId": "9f0b2702803f177f6bac0ab5989d0b8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cake to subscribe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name a customer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email a customer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}